Se refactorizan otros controladores en el manejo de respuestas del servidor

+++ src/main/java/com/tictac/demo/service/MomentoService.java
@@ -19,11 +19,36 @@ public class MomentoService {
     }
     
     public Momento createMomento(Momento momento){
-        return momentoRepository.save(momento);
+        if(momento.getIdHerramienta() == null || momento.getIdHerramienta().toString().trim().isEmpty() ||
+            momento.getNombre() == null || momento.getNombre().trim().isEmpty() ||
+            momento.getDescripcion() == null || momento.getDescripcion().trim().isEmpty()){
+            return null;
+        }else{
+            return momentoRepository.save(momento);
+        }
     }
 
-    public void deleteMomento(Integer id){
-        momentoRepository.deleteById(id);
+    public String updateMomento(Momento momento){
+        if(momentoRepository.existsById(momento.getIdMomento())){
+            Optional<Momento> m = momentoRepository.findById(momento.getIdMomento());
+
+            m.get().setIdHerramienta(momento.getIdHerramienta());
+            m.get().setNombre(momento.getNombre());
+            m.get().setDescripcion(momento.getDescripcion());
+            momentoRepository.save(m.get());
+            return "Ciudad actualizada con éxito";
+        }else{
+            return null;
+        }
+    }
+
+    public String deleteMomento(Integer id){
+        if(momentoRepository.existsById(id)){
+            momentoRepository.deleteById(id);
+            return "Momento eliminado con éxito";
+        }else {
+            return null;
+        }
     }
